requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
langchain
faiss-cpu
langchain-huggingface
sentence-transformers
pypdf